requirements.txt
absl-py==1.0.0
addict==2.4.0
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.9
cycler==0.11.0
Cython==0.29.24
dataclasses @ file:///tmp/build/80754af9/dataclasses_1614363715916/work
decorator==4.4.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.42.0
idna==3.3
imageio==2.13.1
imgaug==0.4.0
importlib-metadata==4.8.2
kiwisolver==1.3.1
Markdown==3.3.6
matplotlib==3.3.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mmcv==1.4.0
networkx==2.5.1
numpy==1.19.5
oauthlib==3.1.1
olefile==0.46
opencv-python==3.4.0.12
packaging==21.3
pickleshare==0.7.5
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.3
pyparsing==3.0.6
pyserial==3.5
python-dateutil==2.8.2
pytorchyolo==1.6.0
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.17.2
scipy==1.5.4
Shapely==1.8.0
six==1.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4.1
terminaltables==3.1.0
tifffile==2020.9.3
torch==1.10.0
torchaudio==0.10.0
torchsummary==1.5.1
torchvision==0.11.1
tqdm==4.62.3
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3==1.26.7
Werkzeug==2.0.2
yapf==0.31.0
zipp==3.6.0
